third_party/WebKit/LayoutTests/fast/css/sticky/sticky-position-valid-after-insertion.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
body {
    margin: 0;
}

.scroller {
    height: 100px;
    width: 200px;
    overflow: scroll;
}

.padding {
    width: 400px;
    height: 500px;
}

.sticky {
    width: 25px;
    height: 25px;
    position: sticky;
    top: 20px;
    left: 50px;
}
</style>

<div class="scroller" id="scroller1">
  <div class="padding" id="padding1"></div>
</div>

<div class="scroller" id="scroller2">
  <div class="padding" id="padding2"></div>
</div>

<script>
if (window.internals) {
  internals.settings.setCSSStickyPositionEnabled(true);
}

test(function() {
    var scroller = document.getElementById('scroller1');
    scroller.scrollTop = 100;
    scroller.scrollLeft = 75;

    var sticky = document.createElement('div');
    sticky.className = 'sticky';
    scroller.insertBefore(sticky, document.getElementById('padding1'));

    assert_equals(sticky.offsetTop, scroller.scrollTop + 20);
    assert_equals(sticky.offsetLeft, scroller.scrollLeft + 50);
}, "offsetTop/offsetLeft should be correct for sticky after insertion");

test(function() {
    var scroller = document.getElementById('scroller2');
    scroller.scrollTop = 100;
    scroller.scrollLeft = 75;

    var sticky = document.createElement('div');
    sticky.className = 'sticky';
    scroller.insertBefore(sticky, document.getElementById('padding2'));

    var sticky_bounds = sticky.getBoundingClientRect();
    var scroller_bounds = scroller.getBoundingClientRect();
    assert_equals(sticky_bounds.top, scroller_bounds.top + 20);
    assert_equals(sticky_bounds.left, scroller_bounds.left + 50);
}, "getBoundingClientRect should be correct for sticky after insertion");

</script>
